+++ api/v1/inc/arbeit.inc.php
@@ -14,13 +14,16 @@ require_once dirname(__DIR__, 1) . "/class/kalender/kalender.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/i18n/i18n.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/benutzer/benutzer.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/auth/auth.arbeit.inc.php";
-require_once dirname(__DIR__, 1) . "/class/pdf/pdf.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/mode/mode.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/kalender/plugins/autodelete.kalender.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/exceptions/exceptions.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/vacation/vacation.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/class/sickness/sickness.arbeit.inc.php";
 require_once $_SERVER["DOCUMENT_ROOT"] . "/vendor/autoload.php";
+
+require_once dirname(__DIR__, 1) . "/class/exports/ExportModule.arbeit.inc.php";
+require_once dirname(__DIR__, 1) . "/class/exports/modules/ExportModuleInterface.em.arbeit.inc.php";
+
 require_once dirname(__DIR__, 1) . "/toil/toil.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/toil/Routes.toil.arbeit.inc.php";
 require_once dirname(__DIR__, 1) . "/toil/Controller.toil.arbeit.inc.php";