+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v7.4
+
+* Exports are now handled differently. You can import your own modules to support more file formats. Read more about this in the `README.md` file.
+* Replaced `Arbeitszeit\PDF` with the new `Arbeitszeit\ExportModule\PDFExportModule` class
+
 ## v7.3.1
 
 * Fixed being unable to toggle easymode